+++ Flow/test/utopia/flow/test/collection/PairTest.scala
@@ -1,7 +1,9 @@
 package utopia.flow.test.collection
 
+import utopia.flow.collection.CollectionExtensions._
 import utopia.flow.collection.immutable.Pair
 import utopia.flow.collection.immutable.Pair.PairOrVectorBuilder
+import utopia.flow.collection.immutable.caching.iterable.LazyPair
 
 /**
   * Tests certain Pair functions
@@ -30,6 +32,16 @@ object PairTest extends App
 	// Tests filter
 	assert(p1.filter { _ > 0 } == p1)
 	assert(p1.filter { _ < 0 } == Vector.empty)
+	
+	// Tests lazy pair
+	val counter = Iterator.iterate(1) { _ + 1 }.pollable
+	val p2 = LazyPair.fill { counter.next() }
+	
+	assert(p2.current.isEmpty)
+	assert(p2.second == 1)
+	assert(counter.poll == 2)
+	assert(p2.current.size == 1)
+	assert(p2.first == 2)
 	
 	println("Success!")
 }